emotion detector refactor 2
