--- emotion_detector/__init__.py
@@ -1,1 +1,0 @@
-from . import emotion_detectionfrom emotion_detection import emotion_detector
